Refactor restructure WrapLabel
